imagebuilder: actually support IB from buildbot
ImageBuilder compiled by buildbot doesn't have any package in the
packages directory. Package needs to be downloaded instead.

This works by calling update to the package manage to download the
remove index and download the file.

Fix missing support for this with APK, by configuring the
--repositories-file option and calling the APK update.

Also move the apk add --initdb to package_index.

If CONFIG_SIGNATURE_CHECK is not enabled, the signature is not checked.

Signed-off-by: Christian Marangi <ansuelsmth@gmail.com>